+++ custom_components/maintenance_supporter/frontend/panel-chunks/object-dialog-TN72QO4Z.js
@@ -1,0 +1,2 @@
+/*! maintenance_supporter frontend 2.71.0 */
+import{a}from"/maintenance_supporter_panelfiles/panel-chunks/chunk-WRWSJ7UC.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-7LDVTUFN.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-TR26N2BJ.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-AASUTC7C.js";export{a as MaintenanceObjectDialog};