+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-HQPIAOLU.js
@@ -1,5 +1,5 @@
 /*! maintenance_supporter frontend 2.71.0 */
-import{a as t}from"./chunk-MKETKKBB.js";function e(o){let r=window;r.customCards=r.customCards||[],r.customCards.some(a=>a.type===o.type)||r.customCards.push(o)}var d=t`
+import{a as t}from"./chunk-TX3XY2US.js";function e(o){let r=window;r.customCards=r.customCards||[],r.customCards.some(a=>a.type===o.type)||r.customCards.push(o)}var d=t`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;